VFXACT-637 Make ziva-vfx-utils Python version agnostic
The Maya 2022 support both Python 2 and Python 3,
the Python scripts in ziva-vfx-utils repo shall support both.
The changes are listed in ZivaVFX confluence pages,
"Python 2 -->Python 3 conversion".

Change some iterative methods, like xrange() to range(),
may affect the performance in Python 2 in some degree,
but that is also a good reason to switch to Python 3 earlier.

+++ zBuilder/ui/utils.py
@@ -28,7 +28,7 @@ def dock_window(dialog_class, *args, **kwargs):
     # now lets get a C++ pointer to it using OpenMaya
     control_widget = mui.MQtUtil.findControl(dialog_class.CONTROL_NAME)
     # conver the C++ pointer to Qt object we can use
-    control_wrap = wrapInstance(long(control_widget), QtWidgets.QWidget)
+    control_wrap = wrapInstance(int(control_widget), QtWidgets.QWidget)
 
     # control_wrap is the widget of the docking window and now we can start working with it:
     control_wrap.setAttribute(QtCore.Qt.WA_DeleteOnClose)
